% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_admin.R
\name{wave_chart_2_ad1}
\alias{wave_chart_2_ad1}
\title{Wykresy w raportach z 1. rundy monitoringu na danych administracyjnych}
\usage{
wave_chart_2_ad1(x, tytul)
}
\arguments{
\item{x}{ramka z danymi do przedstawienia na wykresie}

\item{tytul}{tytuł wykresu w formie ciągu tekstowego lub wartość NULL dla
braku tytułu}
}
\value{

}
\description{
Funkcja rysująca wykres warstwowy na potrzeby raportu. Jako
argument zawierający zbiór danych należy podać ramkę danych będącą wynikiem
działania funkcji \code{tab_wykres_wave_ad1()} lub inną, analogiczną.
}
