% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_admin.R
\name{wyk_pion_skum_5_ad1}
\alias{wyk_pion_skum_5_ad1}
\title{Wykresy w raportach z 1. rundy monitoringu na danych administracyjnych}
\usage{
wyk_pion_skum_5_ad1(x, tytul)
}
\arguments{
\item{x}{ramka z danymi do przedstawienia na wykresie}

\item{tytul}{tytuł wykresu w formie ciągu tekstowego lub wartość NULL dla
braku tytułu}
}
\value{

}
\description{
Funkcja rysująca wykres słupkowy, pionowy, skumulowany na
potrzeby raportu.
}
